+++ CHANGES.rst
@@ -5,6 +5,9 @@ Changelog
 ------------------
 
 - #2458 Allow portal_type as a parameter for api.get_icon function
+- #2434 Add string support for interim fields
+- #2417 Use dtime.to_DT instead of api.to_date
+- #2411 Multiselection with duplicates support for interim fields
 - #2460 Fix keyword missing in AnalysisSpecificationView
 - #2462 Fix manually deselected references in sample add form can not be set anymore
 
@@ -12,6 +15,9 @@ Changelog
 2.5.0 (2024-01-03)
 ------------------
 
+- #2435 Add datetime support for interim fields
+- #2434 Add string support for interim fields
+- #2411 Multiselection with duplicates support for interim fields
 - #2459 Fix default landing page is not considered in clients listing
 - #2456 Remove groups from user add form
 - #2455 Fix users/groups overview batch navigation styling